+++ cmd/aws-sso/process_cmd.go
@@ -29,19 +29,15 @@ import (
 
 type ProcessCmd struct {
 	// AWS Params
-	Arn       string `kong:"short='a',help='ARN of role to assume',xor='arn-1',xor='arn-2',predictor='arn'"`
-	AccountId int64  `kong:"name='account',short='A',help='AWS AccountID of role to assume',xor='arn-1',predictor='accountId'"`
-	Role      string `kong:"short='R',help='Name of AWS Role to assume',xor='arn-2',predictor='role'"`
-	Profile   string `kong:"short='p',help='Name of AWS Profile to assume',xor='arn-1',xor='arn-2',predictor='profile'"`
+	Arn        string `kong:"short='a',help='ARN of role to assume',xor='arn-1',xor='arn-2',predictor='arn'"`
+	AccountId  int64  `kong:"name='account',short='A',help='AWS AccountID of role to assume',xor='arn-1',predictor='accountId'"`
+	Role       string `kong:"short='R',help='Name of AWS Role to assume',xor='arn-2',predictor='role'"`
+	Profile    string `kong:"short='p',help='Name of AWS Profile to assume',xor='arn-1',xor='arn-2',predictor='profile'"`
+	STSRefresh bool   `kong:"help='Force refresh of STS Token Credentials'"`
 }
 
 func (cc *ProcessCmd) Run(ctx *RunContext) error {
 	var err error
-
-	switch ctx.Cli.UrlAction {
-	case "print", "printurl":
-		return fmt.Errorf("unsupported --url-action=print|printurl option")
-	}
 
 	role := ctx.Cli.Process.Role
 	account := ctx.Cli.Process.AccountId
@@ -63,7 +59,7 @@ func (cc *ProcessCmd) Run(ctx *RunContext) error {
 	}
 
 	if role == "" || account == 0 {
-		return fmt.Errorf("Please specify --arn or --account and --role")
+		return fmt.Errorf("please specify --arn or --account and --role")
 	}
 
 	return credentialProcess(ctx, account, role)
@@ -98,7 +94,7 @@ func (cpo *CredentialProcessOutput) Output() (string, error) {
 }
 
 func credentialProcess(ctx *RunContext, accountId int64, role string) error {
-	creds := GetRoleCredentials(ctx, AwsSSO, accountId, role)
+	creds := GetRoleCredentials(ctx, AwsSSO, ctx.Cli.Process.STSRefresh, accountId, role)
 
 	cpo := NewCredentialsProcessOutput(creds)
 	out, err := cpo.Output()