+++ internal/sso/settings.go
@@ -130,7 +130,6 @@ type OverrideSettings struct {
 	DefaultSSO string
 	LogLevel   string
 	LogLines   bool
-	UrlAction  url.Action
 	Threads    int
 }
 
@@ -339,10 +338,6 @@ func (s *Settings) setOverrides(override OverrideSettings) {
 		s.DefaultSSO = override.DefaultSSO
 	}
 
-	if override.UrlAction != "" {
-		s.UrlAction = override.UrlAction
-	}
-
 	if override.Threads > 0 {
 		s.Threads = override.Threads
 	}
@@ -434,7 +429,7 @@ var getExecutable func() (string, error) = func() (string, error) {
 
 // GetAllProfiles returns a map of the ProfileConfig for each SSOConfig.
 // takes the binary path to `open` URL with if set
-func (s *Settings) GetAllProfiles(open url.Action) (*ProfileMap, error) {
+func (s *Settings) GetAllProfiles() (*ProfileMap, error) {
 	profiles := ProfileMap{}
 	var err error
 
@@ -462,7 +457,6 @@ func (s *Settings) GetAllProfiles(open url.Action) (*ProfileMap, error) {
 				BinaryPath:      binaryPath,
 				ConfigVariables: s.ConfigVariables,
 				DefaultRegion:   role.DefaultRegion,
-				Open:            string(open),
 				Profile:         profile,
 				Sso:             ssoName,
 			}